Extract interactive input and output formatting into sub-packages
Splits internal/cli/input.go into internal/cli/input/ package and internal/cli/output.go into internal/cli/output/ package to reduce the monolithic internal/cli package scope.

This addresses the input/output extraction part of issue #324. The vault operations extraction (into internal/vault/operations/) remains for a follow-up.

Closes #324
